+++ packages/integration-sdk-cli/src/bocchi/templates/other/converter.ts.hbs
