+++ packages/integration-sdk-cli/src/bocchi/templates/other/package.json.hbs
