+++ packages/integration-sdk-cli/src/bocchi/templates/steps/fetch-entities.ts.hbs
