+++ packages/integration-sdk-cli/src/bocchi/templates/steps/singleton.ts.hbs
